Amended functions that used nibbles to use bitwise AND, fixed type conflicts

+++ src/assemble.h
@@ -25,7 +25,7 @@ WORD encodeInstruction(char* line);
 /* each encode instruction returns a 32 bit integer instruction */
 
 WORD assembleBranch(enum CondCode condCode, char* target, ADDRESS currentAddress);
-WORD assembleSDT(bool lFlag, REGNUMBER rd, REGNUMBER rn, char* address, int *number);
+WORD assembleSDT(bool lFlag, REGNUMBER rd, REGNUMBER rn, char* address, WORD *number);
 
 WORD assembleMultiply(REGNUMBER rd, REGNUMBER rm, REGNUMBER rs, REGNUMBER rn, bool aFlag);
 WORD assembleDataProc(enum OpCode opCode, REGNUMBER rd, REGNUMBER rn, char* operand2);
